CONTRIB-4145 Switch to new way to specify ignores.
For PHP CodeSniffer 1.4.x and upwards, they way to
specify ignores has changed. It used to be an array
of paths as values. And now it's an array of paths
as keys, with values being "absolute" or "relative".

This commit mimics the conversion performed by the CLI.php,
converting all the to absolute. And seems to work.