CONTRIB-4876 cs: fix the web UI to use the new API.
Errors and warnings are not available anymore from
the CodeSniffer object. Instead, a real report has
to be generated and later build the information
from it.

After balancing if we could use an in-memory custom
report to accumulate coding violations vs. using
some of the existing reports, the later was decided,
using the xml (phpcs internal) format. It allows
quick and easy handling via SimpleXML, keeping the
vast majority of the renderer working in a similar way.