NOBUG: Need to processRuleset() and populateTokenListeners()
phpcs does not do it for us anymore. We need them processed so,
excludes, severity, properties and other bits are considered.

Also need to add the codechecker dir as an standards one, so both
moodle and PHPCompatibility are detected. This seems to be a
requirements for single-sniff runs (like phpunit ones) only. Other
runs (run.php, cli...) work perfectly without needing to specify this.

Finally, add some extra info to be shown when counting expectations
fails. That helps finding the differences a lot.