Add PHPCSAliases.php required for phpcs 3
It seems that there are a good number of old class names
still within the PHPCompatibility standard (to provide
backwards compatibility with phpcs 2). This autoloaded
file makes the magic, allowing their use with phpcs 3.

Modified to be ignored by phpcs itself with phpcs:ignoreFile
and other small changes to make moodlecheck happy.