Temporal workaround for problem including PHPCompatibility
For uses with the bundled phpcs, this is enough and avoids the:

Cannot declare class PHPCompatibility\Sniffs\FunctionUse\RequiredToOptionalFunctionParametersSniff

problem. For using the standard without the bundled phpcs, this still is
a problem (thinking about IDEs, CIs...).

In any case, the problem is reported upstream and eventualy a solution
may arrive. For more info, see:

https://github.com/PHPCompatibility/PHPCompatibility/issues/793