Use annotations in fixture files
It seems that the ruleset.xml properties aren't processed
anymore when running individual Sniffs under testing.

That may have sense, so we are able to cover all different
behaviours instead of just covering the standard one. So
it's assumed that it's a correct feature.

So, now, we have to slightly modify our fixture files to
add phpcs: annotations to explicitly say which properties
to use when running the tests. They should match 100%
the ruleset.xml ones.

I've created https://github.com/squizlabs/PHP_CodeSniffer/issues/3128
to ask if that's intended or no.