Add Github Actions to local_codechecker
Not much to say, using the recommended configuration
and the combinations already used by travis.

Also, fix a couple of warnings to allow checks to pass and
enable a check that was previously commented out in
travis awaiting us to have a new moodle-plugin-ci version
now released.

Finally, remove the 34_STABLE branch (travis & gha),
because it's causing some behat problems, not worth investigating.