Bump highest php version to 8.0 for 311 and master
Only real required change  is to set the new
max_input_vars=5000 requirement. Strictly speaking
only needed for php8, but it doesn't hurt with
older php versions and older moodle branches, so
setting it up always.

About the xmlrpc-beta extension, it's only needed
for php8 runs and only if the plugin uses xmlrpc
for something. If not, it can be omitted without
problem, as far as it's not a moodle requirement
but only a recommendation.

Finally, note that we keep behat & phpunit init and run
disabled, because there are some PHP 8 Warnings that
need to be fixed (see MDL-71622) before enabling them back.

For some reason GHA isn't able to detect those warnings and
everything is passing here, so we continue covered.