Allow all components to be loaded from file instead of calculated
A new phpcs config setting: moodleComponentsListPath

Enables the possibility of loading the list of components from
a file, instead of calculating them from codebase.

This is interesting when the complete codebase is not available,
like happens for example with CiBoT, that only keeps the files
modified in place, to reduce the stuff being checked.

The format of the components file is the same than the
moodle-local_ci/list_valid_components one:

[plugin|subsystem],component_name,component_full_path

Covered with unit tests, will be used in the remote branch checker
(a.k.a. CiBoT).

If the setting is not configured then the standard load of
components from codebase (core_component execution) is performed.