Make the sniff to also detect unexpected MOODLE_INTERNAL checks
Until now, the Sniff was only finding missing MOODLE_INTERNAL checks
(when the file has side effects OR multiple artifacts).

Now, when the check exists and the file does not have side effects
or multiple artifacts, a new MoodleInternalNotNeeded is thrown.

Also, when the old MOODLE_INTERNAL check is detected, a new
MoodleInternalOld warning is thrown.

Covered with tests.

Finally, all the tests related with this Sniff have been moved
to a different test case and all the fixtures moved to new
subdirectory. Part of plans to split the tests completely, with
each sniff having its own testcase.

Fixes #152