Allow @codeCoverageIgnore inline comments
While playing with code coverage, there are some exceptional
places that are not reachable from PHPUnit. So, it's legit to
use the @codeCoverageIgnore annotations, that can be put
both in phpdoc blocks and in inline comments.

This commits enables the later (inline). For phpdoc block we'll
have to enable them in local_moodlecheck.

Covered with tests.