Adding moodle-cs to thirdparty and amend tests
We are adding the moodle-cs copy to the thirdpartylibs.xml file.

That, automatically makes own codechecker and moodlecheck tests
to pass, because that directory is excluded from examinations.

Also, modify the behat tests to stop using the moodle-cs standard
files as fixtures and, instead, be 100% self-contained, using own
files as fixtures.

We also need to move moodle-plugin-ci ignores to tests/fixtures
so they are not processed by CIs.

Finally, remove the base local_codechecker_testcase.php testcase,
because now it's part of moodle-cs.