Make a link @ moodle pointing to new MoodleCS/moodle standard
That should allow IDEs to continue using the old configuration
(pointing to moodle) but using the new one imported from moodle-cs.

I've tested this with vim and it has continued finding and using
the new standard perfectly.

Also, ensure that the "moodle" interim alias is not checked

This alias has been created to allow existing integrations
(with editors, IDEs...) that were pointing to the moodle
directory to specify the standard to continue working.

Now the preferred integration is to use moodle-cs and the
alias will be removed soon. For more information about the
removal see:

https://github.com/moodlehq/moodle-local_codechecker/issues/202

So, in the mean time, to avoid that "moodle" alias to cause
all the code under MoodleCS to be checker (it's not Moodle code!)
we are adding the alias as a dummy thirdpartylibs (the very same
than MoodleCS is).

In one year they will be out, by the issue shared above.