Stop suggesting local_codechecker for composer installs
We already have moodle-cs split into its own repository:

https://github.com/moodlehq/moodle-cs

It can be installed using composer and it's the prefered
method to integrate the standard and PHP_CodeSniffer with
any IDE or tool.

In fact, out own tools (CiBoT and moodle-plugin-ci) have
already switched to use moodle-cs instead of local_codechecker.

So, local_codechecker is just another consumer of moodle-cs
and it's not, anymore, the source for the standard.

This reverts commit 7492e90bfa5179a0a0ce8a0bf121a7c89b6ffdbb.

Also, amended readme instructions to point to moodle-cs.