Allow CRLF files in /tests/fixtures/ directories
In #203 it was proposed to allow CRLF files in Moodle, because
the original CSV specs say that those files should be CRLF.

But in practice, nowadays hardly any file is CRLF and, specifically
for Moodle, we have banned them in core since the night of the times.

And we should continue not allowing them. Hence we are only allowing
those CRLF CSV files within the /tests/fixtures/ directories, that
way if anybody wants to test their CSV loaders with CRLF, they can
have some fixtures at hand.

Everywhere else, they will continue being forbidden.

Fixes #203