Shortcut core's .phpcs.xml by adding our own one
Some problems are happening with core's phpcs.xml
configuration, because it points to the "moodle"
standard and that, as directory, doesn't exist
any more in codechecker (it's an alias now).

But it's an alias that only works under unix
environments, now windows.

So we are adding our own phpcs.xml file, pointing to
the correct standard. That will, automatically, stop
phpcs to continue looking upwards.

See #207 for more detailed information about why core
phpcs.xml is interfering in local_codechecker and,
specifically, breaking under windows.

Fixes #207