Update readme_moodle and libraries
Describe the new steps needed to update the components (basically
to updating moodle-cs), and update the thirdpartylibs.xml entries
to point to the new location in the vendor directory (that also
makes them to be ignored by codechecker itself).

Special note about the phpcompatibility dev version that we
are using and we'll be able to remove it once we switch to
stable releases.

Update phpunit configuration to avoid any coverage within the vendor dir.