Manually remove moodle-cs tests
This step will be needed until this PR:

https://github.com/moodlehq/moodle-cs/pull/69

is merged and released. Then we can get rid of it. Note
the reason is that phpunit finds the tests under vendor
and tries to execute them. As far as moodle-cs is now
just a dependency we don't want it to contain any
developing code (ci, tests, docs...) in its distribution form.